+++ src/Loader/ContentLoader.cs
@@ -20,6 +20,7 @@ namespace NpcAdventure.Loader
         public IDataHelper Data { get; }
         private IContentHelper Assets { get; }
         public ITranslationHelper Translation { get; }
+        public Dictionary<string, string> Translations { get; }
         public string DirectoryPath { get; }
 
         /// <summary>
@@ -37,6 +38,7 @@ namespace NpcAdventure.Loader
             this.contentPacks = helper.ContentPacks;
             this.contentPackManager = contentPackManager ?? throw new ArgumentNullException(nameof(contentPackManager));
             this.monitor = monitor ?? throw new ArgumentNullException(nameof(monitor));
+            this.Translations = new Dictionary<string, string>();
         }
 
         /// <summary>
@@ -47,18 +49,16 @@ namespace NpcAdventure.Loader
         /// <typeparam name="TValue"></typeparam>
         /// <param name="path">Name of asset, like `Strings/Strings` or `Dialogue/Abigail` and etc</param>
         /// <returns>Loaded content of asset</returns>
-        public Dictionary<TKey, TValue> LoadData<TKey, TValue>(string path)
+        public Dictionary<TKey, TValue> LoadData<TKey, TValue>(string path, bool isRequired = true)
         {
             if (!this.TryLoadData<TKey, TValue>(path, out var data))
-                this.monitor.Log($"Cannot load asset: {path}", LogLevel.Error);
+                this.monitor.Log($"Cannot load asset: {path}", isRequired ? LogLevel.Error : LogLevel.Trace);
 
             return data;
         }
 
         private bool TryLoadData<TKey, TValue>(string path, out Dictionary<TKey, TValue> data)
         {
-            string locale = this.Assets.CurrentLocale.ToLower();
-
             // If this content doesn't exists in mod scope, try load them from content packs
             if (!this.HasFile($"assets/{path}.json"))
             {
@@ -67,7 +67,7 @@ namespace NpcAdventure.Loader
 
             data = this.Assets.Load<Dictionary<TKey, TValue>>($"assets/{path}.json");
 
-            this.ApplyTranslation(path, locale, data); // Apply translations                
+            this.ApplyTranslation(path, this.Assets.CurrentLocale, data); // Apply translations                
             this.contentPackManager.Apply(data, path); // Apply content packs
 
             return true;
@@ -101,18 +101,21 @@ namespace NpcAdventure.Loader
             if (string.IsNullOrEmpty(locale))
                 return;
 
-            this.monitor.VerboseLog($"Trying to load localised file `locale/{locale}/{path}.json` for `{path}`, locale `{locale}`");
-            var translatedData = this.Data.ReadJsonFile<Dictionary<TKey, TValue>>($"locale/{locale}/{path}.json");
-
-            if (translatedData == null)
+            if (this.Translations.TryGetValue(locale, out string localeRoot))
             {
-                this.monitor.Log($"No translations for {path} locale {locale}");
-                return;
-                
+                this.monitor.VerboseLog($"Trying to load localised file `{localeRoot}/{path}.json` for `{path}`, locale `{locale}`");
+                var translatedData = this.Data.ReadJsonFile<Dictionary<TKey, TValue>>($"{localeRoot}/{path}.json");
+
+                if (translatedData != null)
+                {
+                    AssetPatchHelper.ApplyPatch(baseData, translatedData);
+                    this.monitor.Log($"Applied mod's translation to `{locale}` for `{path}`.");
+
+                    return;
+                }
             }
 
-            AssetPatchHelper.ApplyPatch(baseData, translatedData);
-            this.monitor.Log($"Applied mod's translation to `{locale}` for `{path}`.");
+            this.monitor.LogOnce($"No translations for {path} locale {locale}", LogLevel.Debug);
         }
 
         /// <summary>